serf 0.7.0 (new formula)
This is Hashicorp's Serf tool, not the Serf library used in SVN.
This has been proposed in homebrew-binary several times, but rejected
for want of a source build:

-   Homebrew/homebrew-binary#236
-   Homebrew/homebrew-binary#110
-   Homebrew/homebrew-binary#98
-   Homebrew/homebrew-binary#26

At one time, there was another formula called serf.rb which built the
SVN library; that library is now built as part of SVN and the formula
was removed some years ago. I propose that it is thus safe to add this
Serf under that name; however, it could be renamed to serfdom.rb (to
match its domain name) or some other name if the maintainers deem it
necessary.

Closes #48520.

Signed-off-by: Dominyk Tiller <dominyktiller@gmail.com>